+++ snapshot-tests-core/src/test/java/de/skuzzle/test/snapshots/impl/FailingSnapshotTests.java
@@ -7,6 +7,7 @@ import java.nio.file.Paths;
 
 import org.junit.jupiter.api.Test;
 import org.opentest4j.AssertionFailedError;
+import org.opentest4j.TestAbortedException;
 
 import de.skuzzle.test.snapshots.EnableSnapshotTests;
 import de.skuzzle.test.snapshots.ForceUpdateSnapshots;
@@ -183,10 +184,10 @@ public class FailingSnapshotTests {
     void testFailBecauseForceUpdateFromAnnotationOnTestClass() throws Throwable {
         frameworkTest
                 .expectTestcase(FailBecauseForceUpdateFromAnnotationOnTestClass.class)
-                .toFailWithExceptionWhich()
-                .isInstanceOf(AssertionError.class)
-                .hasMessage(String.format(
-                        "Snapshots have been updated forcefully.%nRemove '@ForceUpdateSnapshots' annotation from your test class and calls to 'justUpdateSnapshot()' then run the tests again."));
+                .toAllFailWithExceptionWhich(matches -> matches
+                        .isInstanceOf(AssertionError.class)
+                        .hasMessage(String.format(
+                                "Snapshots have been updated forcefully.%nRemove '@ForceUpdateSnapshots' annotation from your test class and calls to 'justUpdateSnapshot()' then run the tests again.")));
     }
 
     @EnableSnapshotTests
@@ -343,10 +344,10 @@ public class FailingSnapshotTests {
     void testFailBecauseInitial() throws Throwable {
         frameworkTest
                 .expectTestcase(FailBecauseInitial.class)
-                .toFailWithExceptionWhich()
-                .isInstanceOf(AssertionError.class)
-                .hasMessage(String.format("Snapshots have been created the first time.%n"
-                        + "Run the test again and you should see it succeed."));
+                .toAllFailWithExceptionWhich(matches -> matches
+                        .isInstanceOf(AssertionError.class)
+                        .hasMessage(String.format("Snapshots have been created the first time.%n"
+                                + "Run the test again and you should see it succeed.")));
     }
 
     @EnableSnapshotTests
@@ -356,6 +357,17 @@ public class FailingSnapshotTests {
         void testWithSnapshot(Snapshot snapshot) throws Throwable {
             MetaTest.assumeMetaTest();
 
+            final SnapshotTestResult snapshotResult = snapshot.assertThat("test").asText().matchesSnapshotText();
+            assertThat(snapshotResult.contextFiles().snapshotFile()).exists();
+            snapshotResult.deleteSnapshot();
+            assertThat(snapshotResult.status()).isEqualTo(SnapshotStatus.CREATED_INITIALLY);
+        }
+
+        @Test
+        void testDisabledAssertionAndInitialAssertion(Snapshot snapshot) throws Throwable {
+            MetaTest.assumeMetaTest();
+
+            snapshot.assertThat("xyz").asText().disabled();
             final SnapshotTestResult snapshotResult = snapshot.assertThat("test").asText().matchesSnapshotText();
             snapshotResult.deleteSnapshot();
             assertThat(snapshotResult.status()).isEqualTo(SnapshotStatus.CREATED_INITIALLY);
@@ -452,6 +464,27 @@ public class FailingSnapshotTests {
 
             snapshot.assertThat("   test   ").as(TextSnapshot.text().withIgnoreWhitespaces(false))
                     .matchesSnapshotStructure();
+        }
+    }
+
+    @Test
+    void testFailWithAssumptionFailed() throws Exception {
+        frameworkTest.expectTestcase(ClassWithDisabledTest.class)
+                .toAllFailWithExceptionWhich(matches -> matches
+                        .isInstanceOf(TestAbortedException.class));
+    }
+
+    @EnableSnapshotTests
+    static class ClassWithDisabledTest {
+        @Test
+        void testWithDisabledAssertion(Snapshot snapshot) throws Exception {
+            snapshot.assertThat("xyz").asText().disabled();
+        }
+
+        @Test
+        void testWithDisabledAndSuccessfulAssertion(Snapshot snapshot) throws Exception {
+            snapshot.assertThat("xyz").asText().disabled();
+            snapshot.assertThat("xyz").asText().matchesSnapshotText();
         }
     }
 }