+++ RELEASE_NOTES.md
@@ -1,3 +1,4 @@
+* [#52](https://github.com/skuzzle/snapshot-tests/issues/52): Test methods that contain at least on `disabled` assertion and o failed assertions will properly be marked as 'skipped' by the test framework
 * Stream line internal implementation
 * Add `ContextFiles` class which groups the paths to all generated files
 * Deprecate `SnapshotTestResult.targetFile()`, `SnapshotTestResult.actualResultFile()` and  `SnapshotTestResult.rawActualResultFile()` in favor of `SnapshotTestResult.contextFiles()`